### 2023-04-12 #5355 @editorialbot
[PRE REVIEW]: PYDAQ: Data Acquisition and Experimental Analysis with Python
**Submitting author:** @samirmartins (<a href="http://orcid.org/0000-0003-1702-8504">Samir Martins</a>)
**Repository:** https://github.com/samirmartins/pydaq
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.3
