+++ paper/paper.bib
@@ -15,7 +15,8 @@
   title={System identification: theory for the user},
   author={Ljung, Lennart},
   year={1987},
-  publisher={Prentice-hall}
+  publisher={Prentice-hall},
+  doi = {10.1016/0005-1098(89)90019-8},
 }
 
 @book{Bil2013,
@@ -24,7 +25,8 @@
   year      = {2013},
   author    = {Billings, S. A.},
   address   = {Chichester},
-  pages     = {574}
+  pages     = {574},
+  doi = {10.1002/9781118535561},
 }
 
 @article{Silva_2018,
@@ -60,7 +62,8 @@ doi={10.1109/EIConRus49466.2020.9038964}
   volume={11},
   pages={100495},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.softx.2020.100495},
 }
 
 @article{WMNL2019,
@@ -70,7 +73,8 @@ doi={10.1109/EIConRus49466.2020.9038964}
   volume={7},
   pages={98228--98237},
   year={2019},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/access.2019.2926057},
 }
 
 @article{MA2016,
@@ -80,7 +84,8 @@ doi={10.1109/EIConRus49466.2020.9038964}
   volume={75},
   pages={607--617},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ymssp.2015.12.031}
 }
 
 @inproceedings{Yang2019,
@@ -114,5 +119,6 @@ doi={10.1109/TIM.2019.2914711}
  journal={Journal of Machine Learning Research},
  volume={12},
  pages={2825--2830},
- year={2011}
+ year={2011},
+ doi={10.48550/arXiv.1201.0490}
 }
