### 2023-04-13 0b95314
Reorganizing code to improve sample time precision
