### 2023-04-13 41d78b0
Updating baud rate considering Arduino UNO
