### 2023-04-13 d57ed8d
Updating baud rate considering Arduino UNO
