### 2023-04-19 ea960c2
# Wait for Arduino and Serial to start up
