### 2023-04-19 e59b0c3
Changing Arduino default baud rate
