### 2023-06-06 #5355 @samirmartins
Dear @pibion and @kyleniemeyer,

I hope this message finds you well. It has been two months since the article submission, and it has not yet been assigned to reviewers. Is there any estimate on when this will happen? Is there anything I can do to help initiate the review process?

Best regards,

Samir
