### 2023-06-07 #5355 @samirmartins
Dear @pibion, thank you for your response. All the reviewers I can recommend have some connection with me (either professionally or personally). If it's possible to suggest these individuals for the review, let me know, and I will tag them in this conversation.

One suggestion is to consider inviting authors of previous articles in JOSS. They might be able to assist with the review.

Kind regards,

Samir Martins
