### 2023-07-11 #5355 @pibion
@nataliakeles, @galessiorob, would you be able to review this submission for JOSS?  The review consists of checking off items that are in a "good" state, and requesting that the repository author improve anything that doesn't yet meet the checklist criteria.
