### 2023-07-11 #5355 @nataliakeles
Yes, I can do it. I will work on this submission.
________________________________
De: pibion ***@***.***>
Enviado: terça-feira, 11 de julho de 2023 00:56
Para: openjournals/joss-reviews ***@***.***>
Cc: Natália Keles ***@***.***>; Mention ***@***.***>
Assunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: PYDAQ: Data Acquisition and Experimental Analysis with Python (Issue #5355)

@nataliakeles<https://github.com/nataliakeles>, @galessiorob<https://github.com/galessiorob>, would you be able to review this submission for JOSS? The review consists of checking off items that are in a "good" state, and requesting that the repository author improve anything that doesn't yet meet the checklist criteria.
