### 2023-07-21 #5662 @nataliakeles
This work is promising so far. It provides an easy-to-use Python library for data acquisition and analysis of experimental developments using NIDAQ or Arduino Microcontrollers. This solution is useful for educational and real-world applications. The text is clear, and there is good documentation on how to use it on GitHub.

However, I suggest the author adopt CapWords convention described in PEP 8 to create class names. The author is referred to the following link https://peps.python.org/pep-0008/#class-names.

An important point is: I don't have a  National Instruments NIDAQ interface to reproduce the examples related to it. What should I do in this case? The code looks fine to me, but I cannot run the examples without the NIDAQ device.
