### 2023-07-21 #5662 @samirmartins
Dear @nataliakeles,

Thank you for your review and comments. In order to follow the CapWords convention (thanks for your suggestion), class names have already been renamed in commits under the branch v0.0.4 and will be released soon. I chose to make this change in the next version to ensure a smooth transition for PYDAQ users.

Regarding testing without a physical NIDAQ device, the National Instruments NIDAQ driver allows users to simulate a virtual NIDAQ board with NI MAX. Using this driver, one can easily simulate both inputs and outputs (https://knowledge.ni.com/KnowledgeArticleDetails?id=kA03q000000x0PxCAI&l=en-US). In fact, PYDAQ was developed using this feature and then tested and used with physical NIDAQ boards that we have in LACOI at UFSJ (acronym for "Control and Instrumentation Laboratory at the Federal University of Sao Joao del-Rei" in Portuguese).

Please let me know if you have any further comments or if I can assist you in any other way.

Best Regards,

Samir
