### 2023-10-24 #5662 @samirmartins
Dear @galessiorob,

Thank you for your review. Regarding the automated tests, you can find them in the pydaq/tests folder, and you can run them using pytest.

@pibion, we now have feedbacks from both reviewers. Is there anything else I can assist with for the continuation and publication of the paper?

Best regards,

Samir
