### 2023-10-25 3359ba7
Adding NI-DAQmx dependence.
