### 2023-12-05 #5662 @samirmartins
Dear @Kevin-Mattheus-Moerman,

As far as I could understand, the checkboxes not marked by the reviewers (they were the same for both reviewers) are those that should not be checked in case of a positive evaluation. Consider, for example, the following:

Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

In other words, in this case, if there are no performance claims, the reviewer should check off this item. Reviewers may (or may not) confirm my interpretation of the text.

Nevertheless, I am entirely at your disposal for any adjustments that may be necessary for the publication of the paper.

Best regards,

Samir
