### 2023-12-13 b98bc2e
Bibliography updates to JOSS paper
