+++ paper/paper.bib
@@ -7,16 +7,18 @@
   number = {54},
   pages = {2384},
   author = {Wilson Rocha Lacerda Junior and Luan Pascoal Costa da Andrade and Samuel Carlos Pessoa Oliveira and Samir Angelo Milani Martins},
-  title = {SysIdentPy: A Python package for System Identification using NARMAX models},
+  title = {{SysIdentPy}: A {Python} package for System Identification using {NARMAX} models},
   journal = {Journal of Open Source Software}
 }
 
 @book{Lju1987,
-  title={System identification: theory for the user},
+  title={System Identification: Theory for the User},
   author={Ljung, Lennart},
-  year={1987},
-  publisher={Prentice-hall},
-  doi = {10.1016/0005-1098(89)90019-8},
+  year={1999},
+  edition={2nd},
+  publisher={Prentice-Hall},
+  address = {Englewood Cliffs, NJ}m
+  isbn = {0-13-656695-2}
 }
 
 @book{Bil2013,
@@ -31,7 +33,6 @@
 
 @article{Silva_2018,
 doi = {10.1088/1361-6404/aae35c},
-url = {https://dx.doi.org/10.1088/1361-6404/aae35c},
 year = {2018},
 month = {oct},
 publisher = {IOP Publishing},
@@ -56,7 +57,7 @@ doi={10.1109/EIConRus49466.2020.9038964}
 }
 
 @article{ayala2020r,
-  title={An R library for nonlinear black-box system identification},
+  title={An {R} library for nonlinear black-box system identification},
   author={Ayala, Helon Vicente Hultmann and Gritti, Marcos Cesar and dos Santos Coelho, Leandro},
   journal={SoftwareX},
   volume={11},
@@ -67,7 +68,7 @@ doi={10.1109/EIConRus49466.2020.9038964}
 }
 
 @article{WMNL2019,
-  title={Control of Hysteretic Systems Through an Analytical Inverse Compensation Based on a NARX Model},
+  title={Control of Hysteretic Systems Through an Analytical Inverse Compensation Based on a {NARX} Model},
   author={Lacerda Junior, Wilson R. and Martins, Samir A. M. and Nepomuceno, Erivelton G. and Lacerda, M{\'a}rcio J.},
   journal={IEEE Access},
   volume={7},
@@ -91,7 +92,7 @@ doi={10.1109/EIConRus49466.2020.9038964}
 @inproceedings{Yang2019,
 author={Yang, Hongxia},
 booktitle={2019 2nd International Conference on Safety Produce Informatization (IICSPI)},
-title={Design and Implementation of Data Acquisition System Based on Scrapy Technology},
+title={Design and Implementation of Data Acquisition System Based on {Scrapy} Technology},
 year={2019},
 volume={},
 number={},
@@ -102,7 +103,7 @@ doi={10.1109/IICSPI48186.2019.9096044}
 @article{Koerner2020,
 author={Koerner, Lucas J. and Caswell, Thomas A. and Allan, Daniel B. and Campbell, Stuart I.},
 journal={IEEE Transactions on Instrumentation and Measurement},
-title={A Python Instrument Control and Data Acquisition Suite for Reproducible Research},
+title={A {Python} Instrument Control and Data Acquisition Suite for Reproducible Research},
 year={2020},
 volume={69},
 number={4},
