### 2023-12-13 #5662 @kyleniemeyer
Hi @samirmartins, I have made some edits to the paper. Can you merge https://github.com/samirmartins/pydaq/pull/8 ?
