### 2023-12-13 #5662 @samirmartins
@kyleniemeyer , all done.

- version number: v0.0.3.1
- There are no releases on Zenodo/figshare/etc

Let me know if you need anything else.

Regards,

> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
>
>     * [x]  Double check authors and affiliations (including ORCIDs)
>
>     * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
>
>     * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.
>
>     * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
>
>     * [x]  Make sure that the license listed for the archive is the same as the software license.
