### 2023-12-14 #5662 @kyleniemeyer
Looks good to me. That DOI warning is erroneous; the associated article is a review of that textbook and does not point to the textbook itself.
