### 2023-12-15 #5662 @samirmartins
> Congratulations @samirmartins on your article's publication in JOSS! Please consider [signing up as a reviewer](https://reviewers.joss.theoj.org/join) if you haven't already.
>
> Many thanks to @galessiorob and @nataliakeles for reviewing this! JOSS wouldn't be possible without your support.

A frase abaixo está correta gramaticalmente?

I'd like to thank you all (@galessiorob , @nataliakeles  and @kyleniemeyer) for your time and effort. @kyleniemeyer, I'm already a JOSS reviewer.

I wish you all the best!

Best regards,

Samir
